pmrep: do not force raw values with CSV output
This is a somewhat unfortunate change but I think it falls into
the better-late-than-never category. For forgotten reasons, CSV
output was forcing raw values (just like archives) but there are
cases when it makes life easier for consumers to see rates rather
than the raw values. Since restoring the old behavior is merely
a matter of adding -r to the command line (as is the case with
other pcp2xxx tools and pmrep/stdout), I think this change in a
major release, although certainly annoying, is still acceptable.

Now pcp2xxx / pmrep csv/stdout are all consistent, for archives
we continue to force storing raw values and that is not changing.