pcp-summary: fix incorrect use of PCP_ORIGIN
We don't want to use this unless -O comes from the command line.

Also -b 1 is safer to use with pminfo in this case of single shot
fetching multiple metrics when using an archive.

This was causing intermittent failures of qa/507 depending on the
way the wind was blowing when pmlogger created the test archive.

Extended qa/507 to use an example archive that showed the failure
case above.