pmie: suppress sleepTight < -1 message if -q on command line
If sleepTight() is called for a task that is more than 1 second
late we emit a warning message as this indicates pmie execution
has been blocked or delayed.

In some QA environments (especially involving VMs), this can happen
and the diagnostic is causing QA failures.  We can now use -q to make
this diagnostic go away (assuming the test is otherwise passing).

Generally pmie can handle these sort of delays, it is just that
the rule evaluation rate will be slower, and expressions involving
any sort of rate calculation will be evaluated over a longer time
period.