libpcp: provide -O/--offset heuristics for local archives
Implement a shorthand notation for the -O/--origin option,
allowing the local system archive directory to be inserted
as an archive name when no other source of metrics has been
supplied.

This was complicated by a quirk of the libpcp getdate.y code
where "today" is (was) an exact synonym for "now".  In this
new mode of operation, we want "today" to mean "the start of
today", instead (& similarly with semantics of "yesterday").
This behaviour been changed in libpcp since its likely to be
a desirable interpreation of these terms in all PCP cases.

New qa test 625 exercises this new -O option behaviour, and
the PCPIntro(1) man page documents it as well.

Resolves github issue #408.