pmie_check & qa/115: fix buglet on lock error path
This has been lurking for ages and causing qa/115 to fail on
some platforms and not on others.

Root cause was a "continue" not inside a loop in _lock() in pmie_check.
On some shells this is a no-op, on others silent badness happens.

Added a bunch of diagnostics in qa/115 that end up in 115.full and
remade the 115.out file.