docs: use OPTIONS section header consistently in man pages
Some pages were using OPTIONS and some COMMAND LINE OPTIONS,
change all using OPTIONS consistently (which is used by Linux
man-pages as well).

There is still some inconsistency with arguments/options
terminology and usage, esp. in PCPIntro(1), but that is not
a case of simple search&replace fix so left for another day.