libpcp: fix label precedence inversion in pmMergeLabelSets
The call to __pmMergeLabelSets() in pmMergeLabelSets had the
parameters around the wrong way, causing higher level labels
to have precedence over the lower levels. This is usually
not obvious from the pminfo -l listing, but showed up with
the (forthcoming new) prometheus PMDA scripting feature that
allows the "hostname" label to be set by a script at the
metric instance level when it fetches metric values from a
remote source (e.g. using a curl script, or whatever).
The instance level hostname label now correctly overrides
the context level hostname label, as it should according to
the PRECEDENCE section in pmLookupLabels(3).

QA for this is included with the updated prometheus PMDA QA tests
coming shortly, see qa/prometheus/scripts/curl_hostname_label.sh