pmda prometheus: add support for scripted configuation
This adds scripts to augment URL configurations for the
prometheus PMDA. Both scripts and URLs can be created in the
$PCP_PMDAS_DIR/prometheus/config.d directory. Note that this
directory has been changed (used to be "urls.d") to reflect
the new functionality. This directory is currently installed
empty (but may gain a README with examples in the near future).

Scripts (or binaries) should be executable and return prometheus
formatted metric data, which is parsed identically to the data
returned by URL files. Files in the configuration directory that
do not end with ".url" and are not executable, are ignored.

Example scripts used for QA may be found in the source directory
qa/prometheus/scripts, along-side the existing example URL files
in qa/prometheus/samples. [this is in the next commit, along with
QA updates].

For hopefully obvious reasons, any script that runs a PCP client
tool to return live PCP data from the local host may only use a
local context (pminfo -L and equiv. for other tools). Remote ingest
and archive sources are not affected.

See the pmdaprometheus(1) man page update for further details,
but basically: scripts should report prometheus data on the
stdout stream and exit successfully in order to have their data
parsed by the PMDA. Anything reported by a script on the stderr
stream will end up in the PMDA log file, which is line buffered.
Note that scripts and URLs are executed and fetched (respectfully)
in parallel threads by the PMDA.

Scripts that retrieve prometheus data from remote host sources
should insert the label "hostname":"remotehost" (for the appropriate
remote hostname) in the prometheus data at the metric instance level.
This is done automatically for URLs because the remost hostname can
be extracted from the URL, discounting any proxy redirection. As
described in pmLookupLabels(3), labels defined at the metric
level will override same named labels further up in the hierarchy
(i.e.  the "hostname" label at the metric instance level will prevail
over the "hostname" label added by pmcd at the context label level).

For security considerations, see a future man page update.