pmlogger: avoid multiple metadata records for the same indom and timestamp
This appears to be a problem that was introduced sometime between the
original code and open sourcing PCP (we don't have source code control
history back that far) but it has only recently surfaced.

If a pmResult contains multiple metrics with the same indom, then
we only need to do the "should we emit an idom metadata record"
test and conditional pmGetInDom() + __pmLogPutInDom() once for
each distinct indom associated with metrics in a pmResult.

This was not always happening and for dynamic indoms (like the proc
metrics) we had the chance of emitting multiple indom metadata records
with the same timestamp and possibly different instances.  The more
proc metrics you log and the faster fork() is being called in the
system the more extreme the problem became.

qa/611 (new, next commit) exposes this issue.

Also added some -Dlogmeta,desperate diagnostics in __localLogGetInDom().

And some editorial/formatting changes to comments.