qa/571: rework filter ... previous .out was wrong
With pmlogger fixes to properly detect indom changes (or not), there
is now an additional indom metadata record that should always have been
there (corner case of the current indom is a proper subset of the indom
at the last pmFetch).

Added output to .full to match timestamps between indom metadata
records and pmResults in the generated archive ... this is not
deterministic due to timing issues between pmlogger and the test
smacking the dynamic indom, so not a .out candidate.