man page changes for NAME section to fix build breakage
The lines after the .SH NAME in man page source files have been (for
more than 40 years) subject to some very fragile and never properly
documented syntax rules.  The \- separator is pretty well known.
The "start a new line creates a new man page" rule is less well known.
And almost a completed secret is that \f3name\f1 is the only safe
way to generate a man page for "name" ... \fBname\fR is the same to
[tng]roff(1), but not the same for the man page generators in the
builds.

This commit addresses the last "rule".  Recent changes in some
of these files were breaking some of the builds.