__pmPrintTimeval() & pmdumplog: increase precision of timestamp output
We ran into a situation where debugging an issue was made much more
confusing by the msec precision user to report timestamps (especially
from an archive log) ... times that looked to be the same were in fact
different by less than 1msec.

__pmPrintTimeval() is used mostly in diagnostics and now reports the
timestamp to usec precision.

pmdumplog is changed to report _all_ times in usec precision.  And as
this changed the formatting of pmdumplog output, some small changes were
made to the format to make it (a) regular, and (b) have less white space
at the start of some lines.

Despite not changing the more user-visible pmPrintStamp() (which still
reports in msec precision), there is a substantial QA fallout from this
change.  Lots of .out files remade, and some tests changed to accommodate
the change pmdumplog reporting format.