pcp2json: add an option to allow to exact types output
In order to maintain the requested precision we need to use the
format() function but this outputs strings, not numbers. round()
outputs numbers but doesn't preserve precision as accurately.

Since there's no one-size-fits-all solution in sight, let's add an
option allowing the user to select the better output mode for a
particular use case.

(Apps should be able to handle numbers as strings since the data
type can be included in output with -x/-X already.)