libpcp & pmdumplog: fix non-interp mode corner cases
As identified in https://github.com/performancecopilot/pcp/issues/418
there was a problem with pmdumplog and -S.

Turns out this is not an issue for interpolation mode, so most
archive uses are immune.

It was a corner case related to the use of the temporal index when
reading serially (forwards or backwards) from someplace not at the
start or end of the archive.

Needed some logic rework in libpcp, and a little attention for
pmdumplog.

Added qa/613 to demonstrate the original issue and a handful of
associated corner cases.