pmdaprometheus: only stat config directories and subdirectories
Only stat the config directories; no need to stat individual config
files - if a change is detected in any of the config directories or
subdirs, then the config files themselves are re-processed for any
changes.

Also checked the overheads - recursive scan and stat on ~ 600 config
directories, three dir levels deep, each containing one or two config
files, takes O(100) ms, so it's pretty light-weight. THis is important
because it's checked on every request to the prometheus pmda.