pmrep: add support for dynamic headers
An early design decision of pmrep was to use a static header due to
the fact that in many cases the availability of instances very rarely
changes (CPUs, disks, NICs, ...). This also helped to keep the code simple.

However, now that new features like instance live filtering, ranking,
and predicates have been added recently, this static header is starting
to limit the usability of these new advanced features. While it would be
possible to first create an archive with pmrep using these features, in
some cases that would be needlessly cumbersome, causing overhead for users.

Thus, there's a need to support dynamic headers with pmrep.

This patch adds dynamic header support for pmrep so that every time there
is a change in available metrics/instances, an updated header is printed.
While this obviously adds some amount of code, with few helpers the
additions are considered bearable and given that these days we have a
huge pmrep/QA test set, we can be certain no regressions are introduced.

All existing QA passing, new QA archive to exercise dynamic headers is
added and a set of extensive test cases are created.  Shell completions
and man pages updated, default configuration and TODO files adjusted.

(Yes, obviously such a dynamic header produces non-standard CSV that
will require post-processing with external tools - if this is not wanted
then it's probably best to write a PCP archive first, and then create
CSV from it. Another alternative is of course simply not to use dynamic
headers with CSV. We allow the user to decide for themselves.)