pmconfig.py: adjust non-leaf metric parsing
It turns out that we can make metricspecs like disk.dm,,,MB to work
with relative ease, and performance is not affected even in extreme
cases (like requesting the entire PMNS tree with "") notably.

It's not perfectly clear yet are there perhaps some hidden corner
cases with this even all the QA passes ok and in general this makes
sense (and allows to get rid of some old cruft). So let's put this
change in but not document it either way (yet) to allow changes in
the future if something surprising comes up.

Few unrelated exception tidy ups / pylint tweaks while at it.