pmrep: fix precision handling for good (again)
It turned out that the "ancient cruft" removed as part of 519de503
was not cruft after all, it is needed for a (corner) case when the
requested precision is higher than the available width. Add a test
case to cover this and put the change partially back in, also note
that this time the QA output change is indeed correct. QA updated.