pmconfig.py: make limit filtering work with predicates
It turns out to be pretty straightforward to make limit filtering
work with predicates.   Do so, and add some QA.  No noticeable
performance penalty (as somewhat expected, the code for other than
predicates+limiting is nothing but one or two additional if's).

Update the earlier rand-pred test archive to be cover this case as
well, add new test cases, fix an off-by-one in previous test cases
to avoid changing their output now that more values is available.