pmrep: keep ranked instances sorted in dynamic headers
When requesting results from pmconfig.py the results are in the order
of instance ids in all cases except when ranking. This doesn't matter
in the case of the static header as only the known instances are read
in the order the header with initially printed. However, when both
the header and the results are dynamic (i.e., ranked), the dynamic
header may get constantly redrawn if an instance value is wavering
and other values are not changing.

We need to sort results ranked results according to instance ids
to avoid this with dynamic headers.