pmdaprometheus: use ctime rather than mtime, and also check stat on individual config files
Noticed pmns changes were not picked-up when an existing config
script is chmod +x. This is because we are only checking mtime
on the directory and subdirs and also because a mode change triggers
a ctime change but not an mtime change. So switch to checking ctime
and also stat all config files as well as directories and subdirs.

This check is called often and it's now a bit more expensive, but not
all that bad - we can still comfortably check hundreds of config files
and subdirs.