Have the PMDA JSON create a default place for json metrics
To make it easier to use the PMDA JSON for dynamically created metrics
it creates a default directory ($PCP_TMP_DIR/json) for other
applications to place their metrics coded as json files in.  PMDA JSON
scans that directory for json files and it will pick up any new
metrics in the default directory automatically.