pmlogger: support PMCD_NAMES_CHANGE protocol for logging dynamic PMNS metrics
PCP GH #372

When a pmFetch returns to pmlogger with the PMCD_NAMES_CHANGE flag set,
check for PMNS changes in non-leaf control blocks originally specified
in the pmlogger config.

This works as follows :

The pmlogger config parser now saves a list of config block names and
associated control parameters (name, status, control and time-delta)
for names that do not resolve to a leaf in the PMNS. The metrics in
this list are potentially dynamic PMNS root nodes, e.g. "prometheus".

When a pmFetch returns with the PMCD_NAMES_CHANGE flag set, pmlogger
traverses the PMNS below each name in the dynamic roots list, and
dynamically amends it's task list with any new metrics discovered.
Any metric names that have been deleted from the PMNS are handled by
existing mechanisms.

The dynamic logging config changes are implemented using the existing
pmlc infrustructure - do_control() has been split into do_control()
and do_control_req(), where the former now decodes a PDU buffer and
then calls the latter, so there are effectively no changes to pmlc
request handling. When pmlogger detects a dynamic PMNS change, it
calls do_control_req() directly to make any required changes to the
logging config and task list.

PMCD_NAMES_CHANGE events are not expected to be frequent, but there is
a fair bit of PMNS lookup roundtrip over-head. AFAICT, this is
unavoidable because the PMCD_NAMES_CHANGE flag is a single bit and
does not indicate which metrics or subtrees have changed or even which
PMDA domain has made dynamic PMNS changes. The increased overheads
can be seen in qa/663. Most other new qa failures are due to new
messages in pmlogger.log and are related to PMCD_NAMES_CHANGE
processing (and thus easily filtered).

We could reduce the overheads in two ways - (a) add a new PMNS API
function to query the dynamic root(s) for a particular PMDA domain,
or (b) add a new 'dynamic' control keyword to the config grammar,
which would be identical to 'mandatory' except the name would specify
the root of a dynamic PMNS subtree. There may be other alternatives
too.

    modified:   src/pmlogger/src/check.c
    modified:   src/pmlogger/src/dopdu.c
    modified:   src/pmlogger/src/fetch.c
    modified:   src/pmlogger/src/gram.y
    modified:   src/pmlogger/src/logger.h
    modified:   src/pmlogger/src/pmlogger.c
    modified:   src/pmlogger/src/ports.c