qa/common.config: change X server in the QA Farm
The host platform was upgraded to Ubuntu 17.10 and I cannot find
a way to configure an X11 server that will accept tcp connections
from remote X11 clients for QA machines that need a remote X server
... this breaks about a dozen tests on about 1/3 of the VMs.

Switch to bozo-laptop that is running a real old LinuxMint distro
where I can get control of the X11 server launch options ... but
X11 freaks on "-" in the hostname part of a $DISPLAY setting, so
use bozo-laptop's IP addr.

Fortunately no one else is exposed to this nonsense.