pmdammv, pmdajson: install a secure default file location
For pcp v4 we switch to installing a PCP_TMP_DIR directory
for both mmv and json instrumented applications.  They are
created pcp:pcp and not-sticky-bit-set by default.

Fix mmv QA tests to use _save_config and _restore_config
so they do the right thing with a pre-installed path.

Also tweak Wills addition to config.json so that the tmpdir
path is inserted during the build, and not Fedora-specific.