pmwebd file service: add/check ETag header for 304 Not Modified results
When pmwebd is used for ordinary file service, it's desirable to allow
web browsers to cache those files.  This patch introduces simple
tagging of these files with the ETag: header (with some readily
available fstat(2) values as keys).  It now sends a 304 result code
instead of a 200 and a whole new copy of unmodified files.