qa: fix multiple misuses of telnet-probe
When testing if a tcp port is in use, the patterns

    echo | telnet-probe localhost <someport>
    echo quit | telnet-probe localhost <someport>

are wrong, as if something unexpected is listening on <someport>
there is a good chance this will hang telnet-probe waiting for
a response to the empty or quit request, or worse cause whoever
is at the end of the port to exit or behave badly.

The pattern

    telnet-probe -c localhost <someport>

should be used as this will exit with out any send-recv traffic
once the connect() succeeds (or fails).

Changes made to the following tests:
    1111 1130 543 589 661 667 912 913 983