pmdaprometheus: only sort the config file list if needed
The config file list only needs to be sorted in a test environment,
where the order of processing end-point sources matters, e.g. for
QA so that we get consistent PMIDs and indom numbers on different
platforms / filesystems.

Outside a test environment, with a large number of source config
files, the sorting is unnecessary and can hurt performance.

So only sort the config list if $PCP_PROMETHEUS_SORT_CONFLIST
is set in the environment. For QA, common.prometheus will set
this.