pmlogger: further refinements to dynamic PMNS handling
GH #372

This builds on the earlier pmlogger dynamic PMNS handling,
see 0c0793672432 which has now been merged into my master
branch.

This is now demonstrated working with qa/1321, subject to
resolving the remaining QA fallout (quite a few failures
in the pmlogger and pmlc groups need to be resolved).

Also refine where the dynamic root list is appended during
config parsing, and lookup each conflig block name to determine
whether it's a leaf-node or a potential dynamic root in the
PMNS.  Unfortunately we can't rely on the return value of
pmTraversePMNS() to determine if a name is a leaf-node or not,
(see GH PCP #441) for details, so the extra lookup is a
work-around - I expect to resolve this soon.

Also wrap the new stderr log messages in pmDebugOptions.log.