pmlogger: fix missing __pmSendError in do_control_req for dynamic PMNS
qa/897 was failing mysteriously because pmlc was unexpectedly
closing it's connection to pmlogger when the test script sent
two consequtive "log mandatory off sample.colour" control requests,
rather than reporting the error, as the qa qualified output was
expecting. Turned out it was a missing __pmSendError from pmlogger
when do_control_req() was failing.

With this commit I think the pmlc / pmlogger control request
processing is now all good after the pmlogger dynamic PMNS
changes. (subject to a verification of a few remaining tests
that are still _notrun on my QA systems).