src/libpcp/src/util.c: remove temp file underneath pmprintf() et al
As part of cleaning up the /tmp and /var/tmp mess we leave behind
after QA, it became evident that pmprintf() was relying on atexit()
to handle the temp file removal and this was not always safe.

In addition the use of a temp file involved a lot of config option
processing and temp dir tests and assorted that is just not necessary.

The message from multiple pmprintf() calls is now assembled in an
in-memory buffer that is expanded in 4096 byte chunks as needed and
freed after pmflush() is finished with it.