libpcp: switch pmGetConfig API to using setenv for valgrind
Mark Wielard has pointed out that we can avoid the valgrind
memory leak warning on initial pmNewContext by using setenv
rather than putenv, and allowing libc to manage the dynamic
memory allocation internally.  This is cosmetic only but is
nice-to-have as valgrind warning suppression is unnecessary
now for PMAPI developers, for this very common case.  QA is
updated to exercise this.

Resolves Red Hat BZ #1546320.