build: fix configure script qmake vs QT_RELEASE corner case
Resolve a situation where QT_RELEASE could still end up not
being set in src/include/builddefs (resulting in cascading
failures later in the build).

When QT_RELEASE is set to specify a preference for Qt5, but
Qt5 is not installed, we want to still fallback to Qt4 (and
certainly not end up with no setting for Qt version at all).