pmrep.conf: add sar -n NFS/NFSD v4 metricsets
It seems better to provide sar-n-NFSv4 / sar-n-NFSDv4 instead
so that we can provide all the needed metrics, avoiding the
nfsclient (optional) PMDA dependency, and also it'll be a bit
easier to customize the metricsets locally if needed.

Those who need NFSv3 metrics (or even NFSv2) can copy/modify
easily the v4 sets to cover v2 or v3.