libpcp: general shell discovery mechanism, Kubernetes pod support
This adds a new discovery mechanism to libpcp service discovery
the pmfind(1) frontend - the "shell" mechanism.  It pipes target
addresses from a directory of shell or other scripts, and probes
them for PCP services in the usual ways.  It is closely based on
the subnet probe mechanism and shares its parallel probing model.

The first shell probe added is pcp-kube-pods(1) which dumps out
a site-configurable list of Kubernetes pod IP addresses to probe.

QA test 1243 is added to exercise everything, uses both valgrind
and helgrind.

As Mark points out, this can be used to remove some untested code
from pmmgr, which may come as a later commit.  pcp-kube-pods has
been written with this replacement and code deletion in mind.