pmdapostgres: fixes and verification to support postgresql v9.6.x
A thorough audit of postgresql 9.6.7 compared to 9.4.5 shows version
9.6 has replaced the pg_stat_activity 'waiting' column with a new column
'wait_event_type' and inserted a new column 'wait_event'. This changes
the order of columns in the pg_stat_activity table, and so some mapping
changes were needed to support 9.6 in the table cherrypick; see related
changes and comments in the refresh_activity function.

The 'waiting' column no longer exists in the table, but the
corresponding PCP metric is now derived as True if the string
value of the new 'wait_event_type' column is not NULL. For
details of the two new columns, see the 9.6 documentation at :
https://www.postgresql.org/docs/9.6/static/monitoring-stats.html#WAIT-EVENT-TABLE

These two new columns have not yet been added as new PCP metrics
(will do so if someone actually asks for them, but they're complicated).
The affected existing metrics: postgresql.stat.activity.waiting and
postgresql.stat.activity.current_query are now verified and working
correctly for 9.6.x. I have also verified with a pgbench TPC-B run and
manually confirmed the reported transaction rates and latency metrics match
up with the pgbench progress reports. This kind of stuff is really difficult
to automatically verify in a QA test - so I didn't try.

I have also audited 9.4 against 9.5 and confirmed there are no table or
column changes affecting any PCP metrics.

No support yet for postgres 10.2, which has been released upstream but not
in the distros yet. That will need another audit and there are very likely
to be more changes needed in the future.

Some QA scripts and table/column audit tools are coming in the next commit.

modified:   src/pmdas/postgresql/pmdapostgresql.1
modified:   src/pmdas/postgresql/pmdapostgresql.pl