pmdapostgresql: automatically reconnect if DB connection is lost
GH #435 - if the postgresql connection is lost, e.g. due to the
service being restarted or whatever, then automatically attempt
to reconnect on the next (and subsequent) fetches until the
connection is re-established. Appropriate messages are sent to
the log.

In addition - install a SIGUSR1 handler to force the PMDA to disconnect.
This is mostly for QA purposes (see qa/1362) to exercise the reconnection
logic without actually bouncing the postgresql service .. since it may
be doing something useful other than QA.

modified:   pmdapostgresql.1
modified:   pmdapostgresql.pl