libpcp.h & pmapi.h: revert some of the libpcp refactoring
There are 3 PM_TEXT_* macros that are needed for internal operations
and sometimes for PMDA developers.  Hiding in them libpcp.h is not
helpful for independent PMDA developers.

Since they are used in conjunction with the other public PM_TEXT_*
macros (they are all bit masks in fact), it seems most logical to
ship all of them together in pmapi.h.