pmlogger_sar_report: write daily summary reports
pmlogger_sar_report is a new cron driven service that writes daily
performance reports, much like sa2(1). It's packaged in the
pcp-zeroconf package (on RPM based systems), and as such is intended
to be installed and run out-of-the-box, with no additional configuration
required. Non RPM packaging is still TODO. By default, pmlogger_sar_report
is invoked as the pcp user by cron at 2am and runs various incantations
of pmrep on yesterdays merged daily archive to produce a report that
is saved by default under $PCP_LOG_DIR/sa/sarXX where XX is yesterdays
day of the month. There are numerous options to change this behaviour,
see the man page and usage message.

Under the hood it just invokes pmrep(1) on each of the :sar-FOO
configs defined in /etc/pcp/pmrep/pmrep.conf. For further details,
see pmlogger_sar_report(1).

New QA test 1180 is coming in the next commit.

        modified:   build/rpm/pcp.spec.in
        new file:   man/man1/pmlogger_sar_report.1
        modified:   src/pmlogger/GNUmakefile
        new file:   src/pmlogger/crontab-sar-report.in
        new file:   src/pmlogger/pmlogger_sar_report.sh