pmlogger_daily: more timely compression changes
There are a bunch of changes here:

- PCP_COMPRESS=foo is an alternate for -X foo
- PCP_COMPRESSAFTER=N is an alternate for -x N
- PCP_COMPRESSREGEX=pat is an alternate for -Y pat

These allow the compression options to be embedded in the pmlogger
control file.

If both the environment variable and the command line argument are
specified for the same option, and they have differrent values,
a warning is emitted and the environment variable "wins".

If PCP_COMPRESSAFTER=0 (or -x 0), then we try to compress all archive
volumes that are not currently being written.

A new -K option (in conjunction with PCP_COMPRESSAFTER=0) does _just_
the compression tasks, and so may be called repeatedly before the
scheduled daily execution of pmlogger_daily.

Extra checks are included to not run pmnewlog if the current pmlogger
process cannot be identified (otherwise we don't know which one to
stop and restart).

Extra checks are included to not compress if the current pmlogger
process, current archive basename and current volume cannot be
identified (otherwise we risk compressing the current volume from
underneath pmlogger).

A new -f option by-passes the extra checks (for QA use only, not
mentioned in the man page).

Safer handling of the pmlogger_daily.prev file.

Some tidying of error and warning messages to improve consistency
and clarity, remove clutter and mention _which_ control file when
context is reported.

man page updates for all of the above.