src/pmlogger/pmlogger_merge.sh: change guards so compressed archives are OK
Needed in conjunction with the $PCP_COMPRESSAFTER=0 changes, as
now if pmlogger is restarted during the day we may end up with some
compressed and some non-compressed archives that need to be
merged.

Note that if this happens, we will need enough disk space to
accommodate one full day's uncompressed archive (in the case of xz
on-the-fly decompression, else twice that much) and one full day's
compressed archives, because we'll have to decompress all the input
archives, merge to produce an uncompressed combined archive, compress
the result and then cleanup the input archives.