pmlogger_daily: change default behaviour for compression
If pmlogconfig -L reports transparent_decompress=true then we will
compress as soon as possible by default ... this is the same as -x 0
or PCP_COMPRESSAFTER=0.  The rationale is that in this environment
we can do on-the-fly decompression, so there is no real reason to
delay compression.

Otherwise compression is never done by default ... this is the
same as -x never or -x forever or PCP_COMPRESSAFTER=never or
PCP_COMPRESSAFTER=forever and matches the previous default for
pmlogger_default, although we shipped crontabs with -x 3 which changed
the effective default.  The -x 3 in the crontabs was removed in a
previous round of commits.

All the PCP_COMPRESSAFTER settings in the control files have been
turned into comments (so users can easily over-ride the defaults
if required).