Debian packaging: make deb-systemd-helper use conditional
In the PCP build we check for older (not real) systemd presence
and choose to not use systemd in the build, which means the
.service files are not installed, which causes deb-systemd-helper
to fail in the recently changed postinst scripts.

This commit adds conditional code in the postinst scripts so
they only use deb-systemd-helper if the corresponding .service
files have been installed, else fallback to the old SysV scheme.