pmdaprometheus: config support for metric filters
GH #453

Add keyword support in URL configs for metric filters. This allows
ingested prometheus metrics to be filtered (included or excluded)
by regex matching of metric names. The syntax is:

FILTER: INCLUDE|EXCLUDE METRIC regex .. to end of line.

Multiple FILTER: lines may be specified, order is important.
The simple rule is the same as for lvm filters - first matching
regex for a particular metric name prevails. The default, if there
are no FILTER: lines in the config, or if no regex matches, is
to include the metric, i.e.

FILTER: INCLUDE METRIC .*

This allows configs like "all metrics except some", and conversely,
"no metrics except these". The latter is expected to be useful for
federated prometheus end-points, which are very verbose.

This is tested in qa/1342. QA updates in the next commit.

Following on from this will be LABEL filters (similar syntax, but
with literal LABEL instead of METRIC in the config entry, along
with additional syntax for optional/mandatory labels. LABEL filters
will allow metrics to be filtered based on whether they have
particular label names and/or value regex matches, as well as
insertion or deletion of individual labels from the incoming
labelset.