pmseries: update for the v2 schema and many other areas
Updates to the pmseries code to sync up the folk working
on this now.

- Implement initial support for native redis time series
  (aka streams) expected to arrive in redis-5.  Probably
  most important there is the much closer mapping to the
  PCP model of metrics and instances within streams, and
  the much-improved memory footprint as a result.
- Switch to using the sds string convenience interfaces
- Switch to using the binary 20-byte sha1 representation
  in-memory (everywhere except in key names, which need
  to be readable and consume relatively little space)
- Drop the use of "hiredis" and switch to a simpler way
  of issuing Redis commands, with less overhead and more
  suited to key extraction for clustering (each key used
  is explicitly provided now, we no longer use sprintf-
  alike high-level interfaces at all).
- Implements pub/sub for newly appearing strings (ie for
  metrics, labels, sources and instance names).
- Update the TODO file for many things now completed and
  several open questions now understood or resolved.

Note: pmseries is (still) not installed (intentionally);
we'll be aiming to get the basic command line interface
in for the next release though, and REST API thereafter.
Initial docs and QA will begin arriving soon as well now
that the schema is (hopefully!) largely stable.  There's
many stability and functionality improvements on the way
first though, so not suited for use beyond those hacking
on it yet.