pcp-pmda-libvirt: rpm packaging changes (spec file changes)
Turns out that on RH systems the required package changed name around
Fedora 27 from libvirt-python to python-libvirt (like the rest of
the world).

But the spec file had Requires: python-libvirt, which did not work
at all on older things like CentOS 7 or RHEL 7 where libvirt-python
is the only package option for the Requires: clause.

This is ugly and will need updating when the name change appears in
the regular RH release series (8?).

The check-vm changes are in commit 0b95f5f that shares this explanation
but does not have the spec file change (too many open windows on too
many machines ...).